# server/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/swagger": {
      "get": {
        "operationId": "AppController_getDevCenterOptions",
        "summary": "Get Dev Center Options",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns dev center options"
          },
          "304": {
            "description": "No changes since last request"
          }
        }
      }
    },
...
    "/api/auth/github/callback": {
      "get": {
        "operationId": "GithubController_githubAuthCallback",
        "summary": "Github OAuth callback",
        "description": "Github OAuth callback URL. This is the URL that Github will redirect to after authentication",
        "parameters": [
          {
            "name": "state",
            "required": false,
            "in": "query",
            "description": "CSRF state token",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "required": true,
            "in": "query",
            "description": "Authorization code from GitHub",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "info": {
    "title": "achievibit-api",
    "description": "![swagger-mode](https://img.shields.io/badge/mode-devcontainer-FF5BF8)\n\nSince this swagger shares the same domain as the app, you can use the same cookie for authentication.\n\nFor the WebSocket API, please visit [achievibit-ws](/api/docs-async)",
    "version": "1.0",
    "contact": {
      "name": "thatkookooguy",
      "url": "https://github.com/thatkookooguy",
      "email": "thatkookooguy@kibibit.io"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT token for authenticated user routes"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "user_token",
        "description": "JWT token for authenticated user routes"
      },
      "gitlab-webhook": {
        "type": "apiKey",
        "name": "x-gitlab-token",
        "in": "header",
        "description": "Secret token for GitLab webhooks validation (usually sent by GitLab)"
      },
      "github-webhook": {
        "type": "apiKey",
        "name": "x-hub-signature-256",
        "in": "header",
        "description": "Secret token for GitHub webhooks validation (usually sent by GitHub)"
      },
      "bitbucket-webhook": {
        "type": "apiKey",
        "name": "x-hub-signature",
        "in": "header",
        "description": "Secret token for Bitbucket webhooks validation (usually sent by Bitbucket)"
      }
    },
...
  }
}